Move rad2deg() and getTouchEventOffsetValues() to timeUtils
